Move public API to simphony.api
