+++ simphony/engine/tests/test_engine_metadata.py
@@ -2,7 +2,7 @@
 import unittest
 from mock import patch
 
-from simphony import CUDS
+from simphony.api import CUDS
 from simphony.cuds.abc_modeling_engine import ABCModelingEngine
 from simphony.engine import ABCEngineExtension,\
     EngineInterface, get_engine_manager, get_supported_engine_names
